requirements.txt
# core and network
httpx==0.27.2
selectolax==0.4.7
fake-useragent==1.5.1
tenacity==8.2.3
curl_cffi==0.7.1

# web and dashboard
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
aiofiles==23.2.1
python-multipart

# database and validation
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
asyncpg==0.30.0

# utilities
loguru==0.7.2
apscheduler==3.10.4

# export
pandas>=2.2.3
openpyxl==3.1.5
xlsxwriter

pgvector==0.3.6
openai==1.55.3
imagehash==4.3.1
Pillow==10.4.0
numpy
aiosmtplib==3.0.2

playwright>=1.49.0
playwright-stealth>=1.0.6